[REL] stock_multi_store to V11. [FIX] Create view files by models. [FIX] Separate classes in diferent files. [FIX] Update Manifest
