+++ stock_multi_store/__init__.py
