+++ stock_multi_store/demo/stock_demo.xml
@@ -1,19 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <data noupdate="1">
+<odoo noupdate="1">
+    <!-- warehouse for rosario -->
+    <record id="stock.warehouse0" model="stock.warehouse">
+        <field name="store_id" ref="base_multi_store.rosario_store"/>
+        <field name="name">YourCompany (Rosario)</field>
+    </record>
 
-        <!-- warehouse for rosario -->
-        <record id="stock.warehouse0" model="stock.warehouse">
-            <field name="store_id" ref="base_multi_store.rosario_store"/>
-            <field name="name">YourCompany (Rosario)</field>
-        </record>
-
-        <record id="warehouse_buenos_aires" model="stock.warehouse">
-            <field name="store_id" ref="base_multi_store.buenos_aires_store"/>
-            <field name="partner_id" ref="base.main_partner"/>
-            <field name="name">Your Company (Bs As)</field>
-            <field name="code">BSAS</field>
-        </record>
-
-    </data>
+    <record id="warehouse_buenos_aires" model="stock.warehouse">
+        <field name="store_id" ref="base_multi_store.buenos_aires_store"/>
+        <field name="partner_id" ref="base.main_partner"/>
+        <field name="name">Your Company (Bs As)</field>
+        <field name="code">BSAS</field>
+    </record>
 </odoo>
