+++ stock_multi_store/models/stock_picking_type.py
@@ -2,8 +2,14 @@
 # For copyright and license notices, see __manifest__.py file in module root
 # directory
 ##############################################################################
-from . import res_store
-from . import stock_picking_type
-from . import stock_picking
-from . import stock_warehouse
+from odoo import models, fields
 
+
+class StockPickingType(models.Model):
+    _inherit = 'stock.picking.type'
+
+    store_id = fields.Many2one(
+        related='warehouse_id.store_id',
+        store=True,
+        readonly=True,
+    )
