+++ stock_multi_store/models/stock_warehouse.py
@@ -15,23 +15,3 @@ class StockWarehouse(models.Model):
         " store, can only see this warehouse records but can not post or"
         " modify any record related to them."
     )
-
-
-class StockPickingType(models.Model):
-    _inherit = 'stock.picking.type'
-
-    store_id = fields.Many2one(
-        related='warehouse_id.store_id',
-        store=True,
-        readonly=True,
-    )
-
-
-class StockPicking(models.Model):
-    _inherit = 'stock.picking'
-
-    store_id = fields.Many2one(
-        related='picking_type_id.store_id',
-        store=True,
-        readonly=True,
-    )
