+++ stock_multi_store/security/multi_store_security.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0"?>
-<odoo>
-    <data noupdate="1">
+<odoo noupdate="1">
 
     <record id="stock_warehouse_store_rule" model="ir.rule">
         <field name="name">Warehouse multi-store</field>
@@ -17,7 +16,7 @@
         <field name="global" eval="True"/>
         <field name="domain_force">['|',('store_id','=',False),('store_id','child_of',[user.store_id.id])]</field>
     </record>
-    
+
     <record id="stock_picking_store_rule" model="ir.rule">
         <field name="name">Stock Picking multi-store</field>
         <field name="model_id" ref="stock.model_stock_picking"/>
@@ -26,6 +25,5 @@
         <field name="domain_force">['|',('store_id','=',False),('store_id','child_of',[user.store_id.id])]</field>
     </record>
 
-    </data>
 </odoo>
