+++ stock_multi_store/views/res_store_view.xml
@@ -1,30 +1,29 @@
 <?xml version="1.0"?>
 <odoo>
-    <data>
+    <!-- Form -->
+    <record id="view_res_store_form" model="ir.ui.view">
+        <field name="name">res.store.form</field>
+        <field name="model">res.store</field>
+        <field name="inherit_id" ref="base_multi_store.view_res_store_form"/>
+        <field name="arch" type="xml">
 
-        <!-- Form -->
-        <record id="view_res_store_form" model="ir.ui.view">
-            <field name="name">res.store.form</field>
-            <field name="model">res.store</field>
-            <field name="inherit_id" ref="base_multi_store.view_res_store_form"/>
-            <field name="arch" type="xml">
-                <div name="buttons">
-                    <button
-                        class="oe_inline oe_stat_button"
-                        name="%(stock.action_warehouse_form)d"
-                        type="action"
-                        context="{'search_default_store_id': active_id}"
-                        icon="fa-home">
-                        <field name="warehouses_count" widget="statinfo"/>
-                    </button>
-                </div>
-                <notebook>
-                    <page string="Warehouses" name="warehouses">
-                        <field name="warehouse_ids"/>
-                    </page>
-                </notebook>
-            </field>
-        </record>
+            <div name="buttons">
+                <button
+                    class="oe_inline oe_stat_button"
+                    name="%(stock.action_warehouse_form)d"
+                    type="action"
+                    context="{'search_default_store_id': active_id}"
+                    icon="fa-home">
+                    <field name="warehouses_count" widget="statinfo"/>
+                </button>
+            </div>
 
-    </data>
+            <notebook>
+                <page string="Warehouses" name="warehouses">
+                    <field name="warehouse_ids"/>
+                </page>
+            </notebook>
+
+        </field>
+    </record>
 </odoo>
