+++ stock_multi_store/views/stock_picking_type_view.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0"?>
+<odoo>
+    <!-- Picking Type Search -->
+    <record id="view_pickingtype_filter" model="ir.ui.view">
+        <field name="name">stock.picking.type.filter</field>
+        <field name="model">stock.picking.type</field>
+        <field name="inherit_id" ref="stock.view_pickingtype_filter"/>
+        <field name="arch" type="xml">
+            <field name="warehouse_id" position="after">
+                <field name="store_id" groups="base_multi_store.group_multi_store"/>
+            </field>
+        </field>
+    </record>
+</odoo>
