+++ stock_multi_store/views/stock_warehouse_view.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0"?>
+<odoo>
+    <!-- Warehouse Form -->
+    <record id="view_warehouse" model="ir.ui.view">
+        <field name="name">stock.warehouse.form</field>
+        <field name="model">stock.warehouse</field>
+        <field name="inherit_id" ref="stock.view_warehouse"/>
+        <field name="arch" type="xml">
+            <field name="company_id" position="after">
+                <field name="store_id" groups="base_multi_store.group_multi_store"/>
+            </field>
+        </field>
+    </record>
+
+    <!-- New Warehouse Search -->
+    <record id="view_stock_warehouse_search" model="ir.ui.view">
+        <field name="name">stock.warehouse.search</field>
+        <field name="model">stock.warehouse</field>
+        <field name="arch" type="xml">
+            <search string="Search Stock_Warehouse">
+                <field name="store_id" groups="base_multi_store.group_multi_store"/>
+            </search>
+        </field>
+    </record>
+</odoo>
